Fix Thunderbird security advisories for patch level releases. #364
Often times, Thunderbird will do a patch level release (eg 91.6.1) that does not
correspond to a Firefox ESR release. (Thunderbird 91.6.1 built against Firefox
91.6.0). When this happens, the security advisories check fails because it is
comparing to the Firefox version found in 'data-esr-versions'.

To get around this without changing the HTML of the advisories page, the id
attribute of the first 'Fixed in Thunderbird XX.X' header element is used.